nirmitee_tally_project/requirements.txt
asgiref==3.8.1
Django==5.1.2
djangorestframework==3.15.2
et_xmlfile==2.0.0
openpyxl==3.1.5
sqlparse==0.5.1
tzdata==2024.2
